requirements.txt
Flask
qrcode[pil]
Werkzeug
gunicorn
supabase-py @ git+https://github.com/supabase/supabase-py.git